Adresacja poprawek z poprzedniego pull request'a:
Zmiana nazwy pakietu
Poprawa literówek w plikach package-info.java
Wyczyszczenie pom.xml
Zmiana serwera z Jetty na Tomcat
