+++ src/main/java/ja/workshop/spring/części/konfiguracja/xml/a_wstrzykiwanie_zależności/silnik/Silnik.java
@@ -1,4 +1,4 @@
-package wrup.spring.a_wstrzykiwanie_zależności.silnik;
+package ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.silnik;
 
 /**
  * Interfejs na fasolki.
