+++ src/main/java/ja/workshop/spring/części/konfiguracja/xml/c_mvc/package-info.java
@@ -7,9 +7,12 @@
  *  <p>Przeanalizuj wszystkie klasy z pakietu, a następnie na podstawie poprzednich dwóch zadań stwórz plik konfiguracyjny <font color="yellow">{@code c_konfiguracja_aplikacji.xml}</font> w pakiecie <font color="yellow">{@code resources/c}</font>,</p>
  *  <p>gdzie zdefiniujesz wszystkie potrzebne ziarna oraz zależności do działania aplikacji.</p>
  *
- *  <p>Do uruchomienia aplikacji wykorzystany zostanie serwer jetty. Jego konfiguracja znajduje się w pliku <font color="yellow">{@code webapp/WEB-INF/web.xml}</font></p>
- *  <p>Aby uruchomić aplikację wpisz w terminalu: <i><b>mvn jetty:run</p></b></i>
+ *  <p>Do uruchomienia aplikacji wykorzystany zostanie serwer tomcat. Jego konfiguracja znajduje się w pliku <font color="yellow">{@code webapp/WEB-INF/web.xml}</font></p>
+ *  <p>Aby uruchomić aplikację wpisz w terminalu:</p>
+ *  <p><b>{@code mvn tomcat7:run}</b>
+ *  <p>lub</p>
+ *  <p><b>{@code mvn tomcat7:run -Dmaven.tomcat.port=8888}</b></p> w przypadku gdy Twój port 8080 jest zajęty.
  *
  * @author Wiktor Rup
  */
-package wrup.spring.c_mvc;
+package ja.workshop.spring.części.konfiguracja.xml.c_mvc;
