+++ src/main/java/ja/workshop/spring/części/konfiguracja/xml/c_mvc/silnik/Silnik.java
@@ -1,0 +1,12 @@
+package ja.workshop.spring.części.konfiguracja.xml.c_mvc.silnik;
+
+/**
+ * Interfejs na ziarna.
+ *
+ * @author Wiktor Rup
+ */
+public interface Silnik {
+
+    String start();
+
+}
