--- src/main/java/wrup/spring/c_mvc/silnik/Silnik.java
@@ -1,10 +1,0 @@
-package wrup.spring.c_mvc.silnik;
-
-/**
- * @author Wiktor Rup
- */
-public interface Silnik {
-
-    String start();
-
-}
