+++ src/main/resources/a/a_konfiguracja.xml
@@ -5,31 +5,31 @@
                            http://www.springframework.org/schema/beans/spring-beans.xsd
                            http://www.springframework.org/schema/context
                            http://www.springframework.org/schema/context/spring-context.xsd">
-    <context:component-scan base-package="wrup.spring.a_wstrzykiwanie_zależności"/>
+    <context:component-scan base-package="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności"/>
     <!-- Definicja ziaren: -->
 
     <!-- 1. Wstrzykiwanie ziarna przez setter -->
 
-    <bean id="silnik" class="wrup.spring.a_wstrzykiwanie_zależności.silnik.OlejNapędowy"/>
-    <bean id="samochód" class="wrup.spring.a_wstrzykiwanie_zależności.Samochód">
+    <bean id="silnik" class="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.silnik.OlejNapędowy"/>
+    <bean id="samochód" class="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.Samochód">
     <property name="silnik" ref="silnik"/>
     </bean>
 
     <!--2. Wstrzykiwanie ziarna przez konstruktor -->
-    <!--<bean id="silnik" class="wrup.spring.a_wstrzykiwanie_zależności.silnik.OlejNapędowy"/>-->
-    <!--<bean id="samochód" class="wrup.spring.a_wstrzykiwanie_zależności.Samochód">-->
+    <!--<bean id="silnik" class="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.silnik.OlejNapędowy"/>-->
+    <!--<bean id="samochód" class="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.Samochód">-->
         <!--<constructor-arg ref="silnik"/>-->
     <!--</bean>-->
 
     <!-- 3. Wstrzykiwanie ziarna fabryką ziaren -->
-    <!--<bean id="fabryczkaFasolekSilników" class="wrup.spring.a_wstrzykiwanie_zależności.silnik.FabryczkaZiarenSilników"/>-->
+    <!--<bean id="fabryczkaFasolekSilników" class="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.silnik.FabryczkaZiarenSilników"/>-->
 
-    <!--<bean id="silnik" class="wrup.spring.a_wstrzykiwanie_zależności.silnik.FabryczkaZiarenSilników"-->
+    <!--<bean id="silnik" class="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.silnik.FabryczkaZiarenSilników"-->
         <!--factory-method="wybierzSilnik" factory-bean="fabryczkaFasolekSilników">-->
         <!--<constructor-arg value="2"/>-->
     <!--</bean>-->
 
-    <!--<bean id="samochód" class="wrup.spring.a_wstrzykiwanie_zależności.Samochód">-->
+    <!--<bean id="samochód" class="ja.workshop.spring.części.konfiguracja.xml.a_wstrzykiwanie_zależności.Samochód">-->
         <!--<property name="silnik" ref="silnik"/>-->
     <!--</bean>-->
