+++ src/main/webapp/WEB-INF/web.xml
@@ -21,6 +21,12 @@
     <servlet-name>KomisKontroler</servlet-name>
     <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
     <init-param>
+        <param-name>contextClass</param-name>
+        <param-value>
+            org.springframework.web.context.support.AnnotationConfigWebApplicationContext
+        </param-value>
+    </init-param>
+    <init-param>
         <param-name>contextConfigLocation</param-name>
         <param-value>classpath:c/c_konfiguracja_web.xml</param-value>
     </init-param>
